chore: 优化用户服务以及基础设施服务的部分逻辑

+++ cdu-snzh/src/main/java/com/snzh/domain/entity/Facility.java
@@ -36,7 +36,7 @@ public class Facility  extends BaseTableData implements Serializable {
     private Long id;
 
     @Schema(description = "设施类型ID")
-    private Integer facilityTypeId;
+    private Long facilityTypeId;
 
     @Schema(description = "设施名称")
     private String name;
